requirements.txt
streamlit>=1.29.0
pandas>=2.2.0
numpy>=1.26.0
faiss-cpu>=1.8.0
Pillow>=10.0.0
pymongo>=4.6.0
dnspython>=2.4.0
python-dotenv>=1.0.0